Add an option to allow non optimal candidates to the approximate_time sync policy
This option allows for low latency between receiving the messages and the execution of the callback.
In the situation of non periodic message streams this becomes quite relevant.

The usual strategy ends up in this situation
 Input A:  a..........b..c.
 Input B:  .A..........B..C
 Output:   ...........a..b.
           ...........A..B.

Wich makes obious that until the message b arrives no callback is called because the optimality can not be proven
but that, in a context of non periodic streams, intoduces unnecessary latency, given that at the second time step
we already have all the messages needed to call the callback.

The flag enables the following situation, at the expense of optimality wich is desirable in some contexts.
 Input A:  a..........b..c.
 Input B:  .A..........B..C
 Output:   .a..........b..c
           .A..........B..C